split gulpfile
